기존의 특정 로그인 버튼을 제네릭하게 변경함.
